Reduced timeout for validator node rounds from 30 to 10 sec

+++ scripts/start.sh
@@ -5,7 +5,6 @@ set -e
 MESSAGE='Usage: init <mode> --monitor
     mode: clean | watch'
 
-
 CURRENT_HOST_IP="$(dig +short myip.opendns.com @resolver1.opendns.com 2>/dev/null || curl -s --retry 2 icanhazip.com)"
 CONSTELLATION_PORT=9000
 
@@ -26,13 +25,13 @@ check_constellation_isStarted(){
 
 NETID=82584648528
 mapfile -t IDENTITY <~/alastria/data/IDENTITY
-GLOBAL_ARGS="--networkid $NETID --identity $IDENTITY --permissioned --rpc --rpcaddr 0.0.0.0 --rpcapi admin,db,eth,debug,miner,net,shh,txpool,personal,web3,quorum,istanbul --rpcport 22000 --port 21000 --istanbul.requesttimeout 30000  --ethstats $IDENTITY:bb98a0b6442386d0cdf8a31b267892c1@52.56.86.239:3000 --verbosity 3 --vmdebug --emitcheckpoints --targetgaslimit 18446744073709551615 --syncmode full "
+GLOBAL_ARGS="--networkid $NETID --identity $IDENTITY --permissioned --rpc --rpcaddr 0.0.0.0 --rpcapi admin,db,eth,debug,miner,net,shh,txpool,personal,web3,quorum,istanbul --rpcport 22000 --port 21000 --istanbul.requesttimeout 10000  --ethstats $IDENTITY:bb98a0b6442386d0cdf8a31b267892c1@52.56.86.239:3000 --verbosity 3 --vmdebug --emitcheckpoints --targetgaslimit 18446744073709551615 --syncmode full "
 
 _TIME=$(date +%Y%m%d%H%M%S)
 
 mapfile -t NODE_TYPE <~/alastria/data/NODE_TYPE
 
-if ( [ "clean" == "$1" ]); then 
+if (  [ ! $# -ne 1 ] && [ "clean" == "$1" ]); then 
     
     echo "Cleaning your node ..."
     rm -rf ~/alastria/logs/quorum_*
@@ -75,12 +74,17 @@ else
 fi
 
 
-if ( [ "--monitor" == "$1" ] ||  [ "--monitor" == "$2" ] ); then
+if ( [ ! $# -ne 1 ] && [ "--monitor" == "$1" ] ); then
     echo "[*] Monitor enabled. Starting monitor..."
     #~/alastria-node/scripts/monitor.sh start 2>&1 > /dev/null
 fi
 
-if ( [ "watch" == "$1" ] )
+if ( [ ! $# -ne 2 ] && [ "--monitor" == "$2" ] ); then
+    echo "[*] Monitor enabled. Starting monitor..."
+    #~/alastria-node/scripts/monitor.sh start 2>&1 > /dev/null
+fi
+
+if ( [ ! $# -ne 1 ] && [ "watch" == "$1" ] )
 then
   ~/alastria-node/scripts/monitor.sh start 2>&1 > /dev/null
   tail -100f ~/alastria/logs/quorum_"${_TIME}".log
